Some css added and files moved around.
