+++ mars-app-spa/src/FCs/dropDownFC.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import Select from "react-select";
-import "./dropDownFC.css";
+import "../Styles/dropDownFC.css";
 
 export const DropDownFC: React.FC<{
   options: { value: any; label: string }[];
