+++ mars-app-spa/src/Styles/SolInput.css
@@ -1,0 +1,3 @@
+.solInputBox {
+    width : 300px
+}
